requirements.txt
streamlit
google-generativeai
langchain
python-dotenv